Reorganize into multiple classes
