+++ Week4/prep-exercises/2-game-of-life/Cell.js
@@ -1,0 +1,40 @@
+/**
+ * @typedef {Object} GridCell
+ * @property {number} x
+ * @property {number} y
+ * @property {boolean} alive
+ * @property {boolean} [nextAlive]
+ */
+
+export default class Cell {
+  static size;
+
+  constructor(x, y) {
+    this.x = x;
+    this.y = y;
+    this.alive = Math.random() > 0.5;
+    this.nextAlive = false;
+  }
+
+  draw(context) {
+    // Draw this background
+    context.fillStyle = '#303030';
+    context.fillRect(
+      this.x * Cell.size,
+      this.y * Cell.size,
+      Cell.size,
+      Cell.size
+    );
+
+    if (this.alive) {
+      // Draw living this inside background
+      context.fillStyle = `rgb(24, 215, 236)`;
+      context.fillRect(
+        this.x * Cell.size + 1,
+        this.y * Cell.size + 1,
+        Cell.size - 2,
+        Cell.size - 2
+      );
+    }
+  }
+}
