+++ Week4/prep-exercises/2-game-of-life/Grid.js
@@ -1,0 +1,77 @@
+import Cell from './Cell.js';
+
+export default class Grid {
+  rows = [];
+
+  constructor(numRows, numColumns, cellSize) {
+    this.numRows = numRows;
+    this.numColumns = numColumns;
+
+    Cell.size = cellSize;
+
+    // Create the grid as a two-dimensional array (i.e. an array of arrays)
+    for (let y = 0; y < numRows; y++) {
+      const row = [];
+      for (let x = 0; x < numColumns; x++) {
+        const cell = new Cell(x, y);
+        row.push(cell);
+      }
+      this.rows.push(row);
+    }
+  }
+
+  isAlive(x, y) {
+    // Out-of-border cells are presumed dead
+    if (x < 0 || x >= this.numColumns || y < 0 || y >= this.numRows) {
+      return 0;
+    }
+
+    const cell = this.rows[y][x];
+    return cell.alive ? 1 : 0;
+  }
+
+  countLivingNeighbors(cell) {
+    const { x, y } = cell;
+    return (
+      this.isAlive(x - 1, y - 1) +
+      this.isAlive(x, y - 1) +
+      this.isAlive(x + 1, y - 1) +
+      this.isAlive(x - 1, y) +
+      this.isAlive(x + 1, y) +
+      this.isAlive(x - 1, y + 1) +
+      this.isAlive(x, y + 1) +
+      this.isAlive(x + 1, y + 1)
+    );
+  }
+
+  forEachCell(callback) {
+    this.rows.forEach((row) => {
+      row.forEach((cell) => callback(cell));
+    });
+  }
+
+  update() {
+    this.forEachCell((cell) => {
+      const numAlive = this.countLivingNeighbors(cell);
+
+      if (numAlive === 2) {
+        // Living cell remains living, dead cell remains dead
+        cell.nextAlive = cell.alive;
+      } else if (numAlive === 3) {
+        // Dead cell becomes living, living cell remains living
+        cell.nextAlive = true;
+      } else {
+        // Living cell dies, dead cell remains dead
+        cell.nextAlive = false;
+      }
+    });
+
+    this.forEachCell((cell) => {
+      cell.alive = cell.nextAlive;
+    });
+  }
+
+  render(context) {
+    this.forEachCell((cell) => cell.draw(context));
+  }
+}
