+++ Week4/prep-exercises/2-game-of-life/README.md
@@ -18,39 +18,59 @@ In the exercise code a new generation of cells replaces the previous one every 2
 
 ## Code walk-through
 
+The JavaScript code is made up of four files, three of which contain JavaScript classes (one class per file) and a file containing a `main()` function.
+
+### `class Cell`
+
+This class represents a single cell.
+
 <!--prettier-ignore-->
-| Methods | Description |
+| Methods  | Description |
 |----------|-------------|
-| `constructor()` | Creates a two-dimensional array (i.e., an array of arrays) that represents a grid of cells that evolve over time. |
-| `createCell()` | (`static` method) Creates a JavaScript object representing a cell with `x` (column number) and `y` (row number) properties and a boolean `aLive` property that is randomly initialized to `true` or `false`. |
+| constructor() | Initializes the cell's `x` and `y` coordinates from arguments passed to the constructor. It randomly sets the initial `alive` boolean state of the cell and initializes it next `alive` state. |
+| draw() | Draws the cell on the canvas. The visual representation depends on whether the cell is alive or dead. |
+
+### `class Grid`
+
+This class manages the game grid, made up of cells.
+
+<!--prettier-ignore-->
+| Methods  | Description |
+|----------|-------------|
+| `constructor()` | Creates a two-dimensional array (i.e., an array of arrays) that represents a grid of cells that evolve over time. It keeps a reference the to the canvas context unto which cells will be drawn. |
 | `forEachCell()` | Executes a callback function for each cell in the two-dimensional grid array, passing the cell as parameter to the callback. |
-| `drawCell()` | Takes a cell object as a parameter and draws the cell on the canvas. The visual representation depends on whether the cell is alive or dead. |
 | `isAlive()` | Determines whether a cell at the given coordinates is alive or dead. The coordinates could potentially be off-grid. Off-grid cells are presumed dead. The function returns one if the given cell is alive or zero if its dead. |
 | `countLivingNeighbors()` | Counts the number of living neighbors for a given cell. Each cell has eight neighbors, some of which may be off-grid if the cell is located at an edge or a corner of the grid. |
-| `updateGrid()` | Iterates through all cells of the grid and computes the new state of each cell by applying the rules of the Game Of Life. |
-| `renderGrid()` | Iterates through all cells of the grid and draws each cell onto the canvas. |
+| `update()` | Iterates through all cells of the grid and computes the new state of each cell by applying the rules of the Game Of Life. |
+| `render()` | Iterates through all cells of the grid and draws each cell onto the canvas. |
+
+### `class Game`
+
+<!--prettier-ignore-->
+| Methods  | Description |
+|----------|-------------|
 | `gameLoop()` | Executes one life cycle of the game (i.e., `updateGrid()` followed by `renderGrid()`) and then reschedules itself to run again after a delay. |
 | `start()` | The `start()` function creates the initial grid, renders it to the web page by calling `renderGrid()` and calls `gameLoop()` to kickstart the game. |
 
-The `main()` function gets a reference to the `canvas` element hard-coded in the `index.html` file and resizes the canvas to the desired size. It then instantiates a GameOfLife object and starts the game engine. The function `main()` itself is executed when the browser has finished loading the page.
+The `main()` function gets a reference to the `canvas` element hard-coded in the `index.html` file and instantiates a `Game` object.
 
-The diagram below visualizes the overall call hierarchy of the various functions. The `main()` function calls `createGame()`, which in turn creates a closure enclosing the `grid` array and a couple of functions that operate on that `grid`. Then, `main()` calls the `start()` function to start the game.
+The diagram below visualizes the overall call hierarchy of the various classes and methods.
 
-The `start()` method creates the initial grid, renders it to the web page by calling `renderGrid()` and calls `gameLoop()` to kickstart the game.
+![Game of Life Call Graph](../../../assets/GameOfLife.png)
 
-The `gameLoop()` method calls `updateGrid()` to update (each cell of) the grid according to the game rules (see above) and the calls `renderGrid()` to render the updated grid to the web page. It then schedules a call to itself using `setTimeout()`. This causes the game to keep evolving the grid according to the game rules every 200ms until the page is closed.
+The `start()` method creates the initial grid and calls `gameLoop()` to kickstart the game.
+
+The `gameLoop()` method calls `grid.update()` to update (each cell of) the grid according to the game rules (see above) and the calls `grid.render()` to render the updated grid to the web page. It then schedules a call to itself using `setTimeout()`. This causes the game to keep evolving the grid according to the game rules every 200ms until the page is closed.
 
 Note: The use of [`window.requestAnimationFrame()`](https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame) is not essential for the functioning of the game but helps to avoid screen flicker.
 
-![Game of Life Call Graph](../../../assets/game-of-life-call-graph.png)
-
 ### Exercise
 
-In the supplied JavaScript code (file: `index.js`) the color of all living cells is a single shade of blue. This is in contrast to the illustration above where living cells have different shades of blue, depending on their life time. Your job is as follows:
+In the supplied JavaScript code the color of all living cells is a single shade of blue. This is in contrast to the illustration above where living cells have different shades of blue, depending on their life time. Your job is as follows:
 
-1. In function `createCell()`, add a numeric `lifeTime` property to the object and assign it the value of one if the cell is initially alive or zero if it is initially dead.
+1. In the constructor of the `Cell` class, add a numeric `lifeTime` property to the object and assign it the value of `1` if the cell is initially alive or `0` if it is initially dead.
 
-2. In function `drawCell()`, replace [`rgb()`](<https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgb()>) with [`rgba()`](<https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgba()>) that adds a fourth parameter indicating `opacity` to the `rgb` value like this:
+2. In `draw` method of the `Cell` class, replace [`rgb()`](<https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgb()>) with [`rgba()`](<https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgba()>) that adds a fourth parameter indicating `opacity` to the `rgb` value like this:
 
    ```js
    context.fillStyle = `rgba(24, 215, 236, ${opacity})`;
@@ -65,7 +85,7 @@ In the supplied JavaScript code (file: `index.js`) the color of all living cells
    |    3     |  0.75   |
    |    4+    |    1    |
 
-3. In function `updateGrid()` add code to update the `lifeTime` value of each cell:
+3. In `update` method of the `Grid` class add code to update the `lifeTime` value of each cell:
 
    - A living cell that remains living should have its `lifeTime` incremented by one.
    - A living cell that dies should have its `lifeTime` reset to zero.
