+++ Week4/prep-exercises/2-game-of-life/index.html
@@ -32,7 +32,7 @@
           >Wikipedia - Conway's Game of Life</a
         >
       </h3>
-      <script src="index.js" type="module"></script>
+      <script src="app.js" type="module"></script>
     </article>
   </body>
 </html>
