+++ eslint.config.js
@@ -1,0 +1,15 @@
+module.exports = [
+  {
+    languageOptions: {
+      parserOptions: {
+        ecmaFeatures: {
+          modules: true,
+        },
+      },
+    },
+    rules: {
+      semi: 'error',
+      'prefer-const': 'error',
+    },
+  },
+];
